Feat:Correção da questão um

+++ src/UmJava.java
@@ -4,7 +4,7 @@ public class UmJava {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int[] num = new int[5];
-        int somaNum = 0;
+        double somaNum = 0;
         for (int i = 0; i < 5; i++) {
             System.out.println("Escreva o " + (i+1) + "° número:");
             num[i] = sc.nextInt();
